ECommerceApi/requirements.txt
asgiref==3.9.1
Brotli==1.1.0
click==8.2.1
dj-database-url==3.0.1
Django==5.2.5
django-filter==25.1
djangorestframework==3.16.1
gunicorn==23.0.0
h11==0.16.0
packaging==25.0
pillow==11.3.0
psycopg2==2.9.10
sqlparse==0.5.3
typing_extensions==4.14.1
uvicorn==0.35.0
whitenoise==6.9.0
